Updated Write-Up
The specification requires that we complete a write-up of the project, specifying end users, possible use cases as well as the design of any databases that may be used in the project. I have been adding to this document steadily for the past year and only just had the genius idea of storing it with the rest of this project.